+++ common/lib/dependabot/git_commit_checker.rb
@@ -267,7 +267,9 @@ module Dependabot
 
     def dependency_source_details
       sources =
-        dependency.requirements.map { |r| r.fetch(:source) }.uniq.compact
+        dependency.requirements.
+        map { |requirement| requirement.fetch(:source) }.uniq.compact.
+        select { |source| source[:type] == "git" }
 
       return sources.first if sources.count <= 1