+++ terraform/spec/dependabot/terraform/file_parser_spec.rb
@@ -12,10 +12,7 @@ RSpec.describe Dependabot::Terraform::FileParser do
   subject(:parser) do
     described_class.new(
       dependency_files: files,
-      source: source,
-      options: {
-        terraform_hcl2: PackageManagerHelper.use_terraform_hcl2?
-      }
+      source: source
     )
   end
 
@@ -23,7 +20,7 @@ RSpec.describe Dependabot::Terraform::FileParser do
   let(:source) { Dependabot::Source.new(provider: "github", repo: "gocardless/bump", directory: "/") }
 
   describe "#parse" do
-    subject { parser.parse }
+    subject(:dependencies) { parser.parse }
 
     context "with an invalid registry source" do
       let(:files) { project_dependency_files("invalid_registry") }
@@ -38,14 +35,7 @@ RSpec.describe Dependabot::Terraform::FileParser do
     context "with an unparseable source" do
       let(:files) { project_dependency_files("unparseable") }
 
-      it "raises an error for hcl1", :hcl1_only do
-        expect { subject }.to raise_error(Dependabot::DependencyFileNotParseable) do |boom|
-          expect(boom.file_path).to eq("/main.tf")
-          expect(boom.message).to eq("unable to parse HCL: object expected closing RBRACE got: EOF")
-        end
-      end
-
-      it "raises an error for hcl2", :hcl2_only do
+      it "raises an error" do
         expect { subject }.to raise_error(Dependabot::DependencyFileNotParseable) do |boom|
           expect(boom.message).to eq(
             "Failed to convert file: parse config: [:18,1-1: Argument or block definition required; " \
@@ -138,8 +128,7 @@ RSpec.describe Dependabot::Terraform::FileParser do
     end
 
     context "with git sources" do
-      let(:files) { project_dependency_files("git_tags") }
-
+      let(:files) { project_dependency_files("git_tags_011") }
       specify { expect(subject.length).to eq(6) }
       specify { expect(subject).to all(be_a(Dependabot::Dependency)) }
 
@@ -240,33 +229,10 @@ RSpec.describe Dependabot::Terraform::FileParser do
       end
     end
 
-    context "with a terragrunt file", :hcl1_only do
-      let(:files) { project_dependency_files("terragrunt") }
-
-      specify { expect(subject.length).to eq(1) }
-      specify { expect(subject).to all(be_a(Dependabot::Dependency)) }
-
-      it "has the right details for the first dependency" do
-        expect(subject[0].name).to eq("gruntwork-io/modules-example")
-        expect(subject[0].version).to eq("0.0.2")
-        expect(subject[0].requirements).to eq([{
-          requirement: nil,
-          groups: [],
-          file: "main.tfvars",
-          source: {
-            type: "git",
-            url: "git@github.com:gruntwork-io/modules-example.git",
-            branch: nil,
-            ref: "v0.0.2"
-          }
-        }])
-      end
-    end
-
     context "hcl2 files" do
       let(:files) { project_dependency_files("hcl2") }
 
-      it "has the right source for the dependency", :hcl2_only do
+      it "has the right source for the dependency" do
         expect(subject[0].requirements).to eq([{
           requirement: nil,
           groups: [],
@@ -275,13 +241,307 @@ RSpec.describe Dependabot::Terraform::FileParser do
             type: "git",
             url: "git@github.com:cloudposse/terraform-aws-jenkins.git",
             branch: nil,
-            ref: "0.4.0"
+            ref: "0.4.1"
           }
         }])
       end
 
-      it "fails to parse hcl2 files without the flag set", :hcl1_only do
-        expect { subject }.to raise_error(Dependabot::DependencyFileNotParseable)
+      context "with git sources" do
+        let(:files) { project_dependency_files("git_tags_012") }
+
+        specify { expect(subject.length).to eq(6) }
+        specify { expect(subject).to all(be_a(Dependabot::Dependency)) }
+
+        it "has the right details for the first dependency (which uses git:: with a tag)" do
+          dependency = subject.find { |x| x.name == "origin_label" }
+          expect(dependency).to_not be_nil
+          expect(dependency.version).to eq("0.3.7")
+          expect(dependency.requirements).to match_array([{
+            requirement: nil,
+            groups: [],
+            file: "main.tf",
+            source: {
+              type: "git",
+              url: "https://github.com/cloudposse/terraform-null-label.git",
+              branch: nil,
+              ref: "tags/0.3.7"
+            }
+          }])
+        end
+
+        it "has the right details for the second dependency (which uses github.com with a tag)" do
+          dependency = subject.find { |x| x.name == "logs" }
+          expect(dependency).to_not be_nil
+          expect(dependency.version).to eq("0.2.2")
+          expect(dependency.requirements).to match_array([{
+            requirement: nil,
+            groups: [],
+            file: "main.tf",
+            source: {
+              type: "git",
+              url: "https://github.com/cloudposse/terraform-aws-s3-log-storage.git",
+              branch: nil,
+              ref: "tags/0.2.2"
+            }
+          }])
+        end
+
+        it "has the right details for the third dependency (which uses bitbucket.org with no tag)" do
+          dependency = subject.find { |x| x.name == "distribution_label" }
+          expect(dependency).to_not be_nil
+          expect(dependency.version).to be_nil
+          expect(dependency.requirements).to eq([{
+            requirement: nil,
+            groups: [],
+            file: "main.tf",
+            source: {
+              type: "git",
+              url: "https://bitbucket.org/cloudposse/terraform-null-label.git",
+              branch: nil,
+              ref: nil
+            }
+          }])
+        end
+
+        it "has the right details the fourth dependency (which has a subdirectory and a tag)" do
+          dependency = subject.find { |x| x.name == "dns" }
+          expect(dependency).to_not be_nil
+          expect(dependency.version).to eq("0.2.5")
+          expect(dependency.requirements).to eq([{
+            requirement: nil,
+            groups: [],
+            file: "main.tf",
+            source: {
+              type: "git",
+              url: "https://github.com/cloudposse/terraform-aws-route53-cluster-zone.git",
+              branch: nil,
+              ref: "tags/0.2.5"
+            }
+          }])
+        end
+
+        it "has the right details the fifth dependency)" do
+          dependency = subject.find { |x| x.name == "duplicate_label" }
+          expect(dependency).to_not be_nil
+          expect(dependency.version).to eq("0.3.7")
+          expect(dependency.requirements).to eq([{
+            requirement: nil,
+            groups: [],
+            file: "main.tf",
+            source: {
+              type: "git",
+              url: "https://github.com/cloudposse/terraform-null-label.git",
+              branch: nil,
+              ref: "tags/0.3.7"
+            }
+          }])
+        end
+
+        it "has the right details for the sixth dependency (which uses git@github.com)" do
+          dependency = subject.find { |x| x.name == "github_ssh_without_protocol" }
+          expect(dependency).to_not be_nil
+          expect(dependency.version).to eq("0.4.0")
+          expect(dependency.requirements).to eq([{
+            requirement: nil,
+            groups: [],
+            file: "main.tf",
+            source: {
+              type: "git",
+              url: "git@github.com:cloudposse/terraform-aws-jenkins.git",
+              ref: "tags/0.4.0",
+              branch: nil
+            }
+          }])
+        end
+      end
+
+      context "with registry sources" do
+        let(:files) { project_dependency_files("registry_012") }
+
+        its(:length) { is_expected.to eq(5) }
+
+        describe "default registry with version" do
+          subject(:dependency) { dependencies.find { |d| d.name == "hashicorp/consul/aws" } }
+          let(:expected_requirements) do
+            [{
+              requirement: "0.1.0",
+              groups: [],
+              file: "main.tf",
+              source: {
+                type: "registry",
+                registry_hostname: "registry.terraform.io",
+                module_identifier: "hashicorp/consul/aws"
+              }
+            }]
+          end
+
+          it "has the right details" do
+            expect(dependency).to be_a(Dependabot::Dependency)
+            expect(dependency.name).to eq("hashicorp/consul/aws")
+            expect(dependency.version).to eq("0.1.0")
+            expect(dependency.requirements).to eq(expected_requirements)
+          end
+        end
+
+        describe "default registry with no version" do
+          subject(:dependency) { dependencies.find { |d| d.name == "devops-workflow/members/github" } }
+          let(:expected_requirements) do
+            [{
+              requirement: nil,
+              groups: [],
+              file: "main.tf",
+              source: {
+                type: "registry",
+                registry_hostname: "registry.terraform.io",
+                module_identifier: "devops-workflow/members/github"
+              }
+            }]
+          end
+
+          it "has the right details" do
+            expect(dependency).to be_a(Dependabot::Dependency)
+            expect(dependency.name).to eq("devops-workflow/members/github")
+            expect(dependency.version).to be_nil
+            expect(dependency.requirements).to eq(expected_requirements)
+          end
+        end
+
+        describe "the third dependency (default registry with a sub-directory)" do
+          subject(:dependency) { dependencies.find { |d| d.name == "mongodb/ecs-task-definition/aws" } }
+          let(:expected_requirements) do
+            [{
+              requirement: nil,
+              groups: [],
+              file: "main.tf",
+              source: {
+                type: "registry",
+                registry_hostname: "registry.terraform.io",
+                module_identifier: "mongodb/ecs-task-definition/aws"
+              }
+            }]
+          end
+
+          it "has the right details" do
+            expect(dependency).to be_a(Dependabot::Dependency)
+            expect(dependency.name).to eq("mongodb/ecs-task-definition/aws")
+            expect(dependency.version).to be_nil
+            expect(dependency.requirements).to eq(expected_requirements)
+          end
+        end
+
+        describe "the fourth dependency (default registry with version req)" do
+          subject(:dependency) { dependencies.find { |d| d.name == "terraform-aws-modules/rds/aws" } }
+          let(:expected_requirements) do
+            [{
+              requirement: "~> 1.0.0",
+              groups: [],
+              file: "main.tf",
+              source: {
+                type: "registry",
+                registry_hostname: "registry.terraform.io",
+                module_identifier: "terraform-aws-modules/rds/aws"
+              }
+            }]
+          end
+
+          it "has the right details" do
+            expect(dependency).to be_a(Dependabot::Dependency)
+            expect(dependency.name).to eq("terraform-aws-modules/rds/aws")
+            expect(dependency.version).to be_nil
+            expect(dependency.requirements).to eq(expected_requirements)
+          end
+        end
+
+        describe "the fifth dependency (private registry with version)" do
+          subject(:dependency) { dependencies.find { |d| d.name == "example_corp/vpc/aws" } }
+          let(:expected_requirements) do
+            [{
+              requirement: "0.9.3",
+              groups: [],
+              file: "main.tf",
+              source: {
+                type: "registry",
+                registry_hostname: "app.terraform.io",
+                module_identifier: "example_corp/vpc/aws"
+              }
+            }]
+          end
+
+          it "has the right details" do
+            expect(dependency).to be_a(Dependabot::Dependency)
+            expect(dependency.name).to eq("example_corp/vpc/aws")
+            expect(dependency.version).to eq("0.9.3")
+            expect(dependency.requirements).to eq(expected_requirements)
+          end
+        end
+      end
+
+      context "with terragrunt files" do
+        let(:files) { project_dependency_files("terragrunt_hcl") }
+
+        specify { expect(subject.length).to eq(1) }
+        specify { expect(subject).to all(be_a(Dependabot::Dependency)) }
+
+        it "has the right details for the first dependency" do
+          expect(subject[0].name).to eq("gruntwork-io/modules-example")
+          expect(subject[0].version).to eq("0.0.2")
+          expect(subject[0].requirements).to eq([{
+            requirement: nil,
+            groups: [],
+            file: "terragrunt.hcl",
+            source: {
+              type: "git",
+              url: "git@github.com:gruntwork-io/modules-example.git",
+              branch: nil,
+              ref: "v0.0.2"
+            }
+          }])
+        end
+      end
+    end
+
+    context "terraform.lock.hcl files" do
+      let(:files) { project_dependency_files("terraform_lock_only") }
+
+      it "does not attempt to parse the lockfile" do
+        expect { dependencies.length }.to raise_error(RuntimeError) do |error|
+          expect(error.message).to eq("No Terraform configuration file!")
+        end
+      end
+    end
+
+    context "with a required provider" do
+      let(:files) { project_dependency_files("registry_provider") }
+
+      it "has the right details" do
+        dependency = dependencies.first
+
+        expect(dependency.name).to eq("hashicorp/aws")
+        expect(dependency.version).to eq("0.1.0")
+      end
+    end
+
+    context "with a required provider block with multiple versions" do
+      let(:files) { project_dependency_files("registry_provider_compound_local_name") }
+
+      it "has the right details" do
+        hashicorp = dependencies.find { |d| d.name == "hashicorp/http" }
+        mycorp = dependencies.find { |d| d.name == "mycorp/http" }
+
+        expect(hashicorp.version).to eq("2.0")
+        expect(mycorp.version).to eq("1.0")
+      end
+    end
+
+    context "with a toplevel provider" do
+      let(:files) { project_dependency_files("provider") }
+
+      it "does not find the details" do
+        # This feature is deprecated as documented here:
+        # https://www.terraform.io/docs/language/providers/configuration.html#version-an-older-way-to-manage-provider-versions
+        # So dependabot does not support it. This test is here for
+        # documentatio-sake.
+        expect(dependencies.count).to eq(0)
       end
     end
   end